+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Redesign work with sqlite
 - Added src/third-party/easyzip/ https://github.com/winsoft666/easyzip (for replacement quazip-0.7.3
 - Redesign zipping to easyzip and removed quazip-0.7.3
+- Added generate models for pm.py
 
 ## [v0.2.52] - 2024-07-21 (2024 Jule 21)