+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to fhq-server project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [v0.2.52] - ??
+
+
+
 ## [v0.2.51] - 2023-01-03 (2023 Jan 3)
 
 ### Added